requirements.txt
streamlit==1.41.1
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-community==0.3.14
faiss-cpu==1.9.0.post1
pypdf==5.1.0
python-dotenv==1.0.1
